refactor(splice): move _cat_length helpers to _splice module
Relocates _cat_length and _cat_length_inner from _dataset/_impl.py into a
new _dataset/_splice.py module. No behavior change; updates the test import
accordingly and removes now-unused awkward layout imports from _impl.py.
